Update with GeoNode Master

+++ geonode/layers/models.py
@@ -105,7 +105,7 @@ class Layer(ResourceBase):
     styles = models.ManyToManyField(Style, related_name='layer_styles')
 
     def update_thumbnail(self, save=True):
-        self.save_thumbnail(self._thumbnail_url(width=198, height=98), save)
+        self.save_thumbnail(self._thumbnail_url(width=200, height=150), save)
 
 
     def _render_thumbnail(self, spec):
@@ -127,12 +127,12 @@ class Layer(ResourceBase):
         if height is not None:
             params['height'] = height
 
-        # Avoid usring urllib.urlencode here because it breaks the url.
+        # Avoid using urllib.urlencode here because it breaks the url.
         # commas and slashes in values get encoded and then cause trouble
         # with the WMS parser.
         p = "&".join("%s=%s"%item for item in params.items())
 
-        return ogc_server_settings.LOCATION + "wms/reflect?" + p
+        return ogc_server_settings.public_url + "wms/reflect?" + p
 
 
     def verify(self):
@@ -208,7 +208,7 @@ class Layer(ResourceBase):
             user = User.objects.get(username=username)
             self.set_user_level(user, self.LEVEL_NONE)
 
-        # assign owner admin privs
+        # assign owner admin privileges
         if self.owner:
             self.set_user_level(self.owner, self.LEVEL_ADMIN)
 
@@ -275,7 +275,9 @@ def geoserver_pre_delete(instance, sender, **kwargs):
     """
     ct = ContentType.objects.get_for_model(instance)
     OverallRating.objects.filter(content_type = ct, object_id = instance.id).delete()
-    cascading_delete(Layer.objects.gs_catalog, instance.typename)
+    #cascading_delete should only be called if ogc_server_settings.BACKEND_WRITE_ENABLED == True
+    if getattr(ogc_server_settings,"BACKEND_WRITE_ENABLED", True):
+        cascading_delete(Layer.objects.gs_catalog, instance.typename)
 
 
 def pre_save_layer(instance, sender, **kwargs):
@@ -314,7 +316,7 @@ def post_delete_layer(instance, sender, **kwargs):
     MapLayer.objects.filter(name=instance.typename).delete()
     logger.debug("Going to delete the default style for [%s]", instance.typename.encode('utf-8'))
 
-    if Layer.objects.filter(default_style__id=instance.default_style.id).count() == 0:
+    if instance.default_style and Layer.objects.filter(default_style__id=instance.default_style.id).count() == 0:
         instance.default_style.delete()
 
 def geoserver_pre_save(instance, sender, **kwargs):
@@ -344,7 +346,7 @@ def geoserver_pre_save(instance, sender, **kwargs):
         return
 
     # If there is no resource returned it could mean one of two things:
-    # a) There is a syncronization problem in geoserver
+    # a) There is a synchronization problem in geoserver
     # b) The unit tests are running and another geoserver is running in the
     # background.
     # For both cases it is sensible to stop processing the layer
@@ -362,7 +364,9 @@ def geoserver_pre_save(instance, sender, **kwargs):
         metadata_links.append((link.name, link.mime, link.url))
 
     gs_resource.metadata_links = metadata_links
-    gs_catalog.save(gs_resource)
+    #gs_resource should only be called if ogc_server_settings.BACKEND_WRITE_ENABLED == True
+    if getattr(ogc_server_settings,"BACKEND_WRITE_ENABLED", True):
+        gs_catalog.save(gs_resource)
 
     gs_layer = gs_catalog.get_layer(instance.name)
 
@@ -370,7 +374,9 @@ def geoserver_pre_save(instance, sender, **kwargs):
         gs_layer.attribution = str(instance.poc.user)
         profile = Profile.objects.get(user=instance.poc.user)
         gs_layer.attribution_link = settings.SITEURL[:-1] + profile.get_absolute_url()
-        gs_catalog.save(gs_layer)
+        #gs_layer should only be called if ogc_server_settings.BACKEND_WRITE_ENABLED == True
+        if getattr(ogc_server_settings,"BACKEND_WRITE_ENABLED", True):
+            gs_catalog.save(gs_layer)
 
     """Get information from geoserver.
 
@@ -401,11 +407,11 @@ def geoserver_pre_save(instance, sender, **kwargs):
 def geoserver_post_save(instance, sender, **kwargs):
     """Save keywords to GeoServer
 
-       The way keywords are implemented require the layer
+       The way keywords are implemented requires the layer
        to be saved to the database before accessing them.
     """
-    url = "%srest" % settings.OGC_SERVER['default']['LOCATION']
-
+    url = ogc_server_settings.rest
+    
     try:
         gs_catalog = Catalog(url, _user, _password)
         gs_resource = gs_catalog.get_resource(instance.name)
@@ -419,7 +425,7 @@ def geoserver_post_save(instance, sender, **kwargs):
         return
 
     # If there is no resource returned it could mean one of two things:
-    # a) There is a syncronization problem in geoserver
+    # a) There is a synchronization problem in geoserver
     # b) The unit tests are running and another geoserver is running in the
     # background.
     # For both cases it is sensible to stop processing the layer
@@ -428,7 +434,9 @@ def geoserver_post_save(instance, sender, **kwargs):
         return
 
     gs_resource.keywords = instance.keyword_list()
-    gs_catalog.save(gs_resource)
+    #gs_resource should only be called if ogc_server_settings.BACKEND_WRITE_ENABLED == True
+    if getattr(ogc_server_settings,"BACKEND_WRITE_ENABLED", True):
+        gs_catalog.save(gs_resource)
 
     bbox = gs_resource.latlon_bbox
     dx = float(bbox[1]) - float(bbox[0])
@@ -441,7 +449,7 @@ def geoserver_post_save(instance, sender, **kwargs):
 
     # Set download links for WMS, WCS or WFS and KML
 
-    links = wms_links(ogc_server_settings.LOCATION + 'wms?',
+    links = wms_links(ogc_server_settings.public_url + 'wms?',
                     instance.typename.encode('utf-8'), instance.bbox_string,
                     instance.srid, height, width)
 
@@ -457,7 +465,7 @@ def geoserver_post_save(instance, sender, **kwargs):
                         )
 
     if instance.storeType == "dataStore":
-        links = wfs_links(ogc_server_settings.LOCATION + 'wfs?', instance.typename.encode('utf-8'))
+        links = wfs_links(ogc_server_settings.public_url + 'wfs?', instance.typename.encode('utf-8'))
         for ext, name, mime, wfs_url in links:
             Link.objects.get_or_create(resource= instance.resourcebase_ptr,
                             url=wfs_url,
@@ -479,7 +487,7 @@ def geoserver_post_save(instance, sender, **kwargs):
         permissions['authenticated'] = instance.get_gen_level(AUTHENTICATED_USERS)
         instance.set_gen_level(ANONYMOUS_USERS,'layer_readonly')
 
-        links = wcs_links(ogc_server_settings.LOCATION + 'wcs?', instance.typename.encode('utf-8'),
+        links = wcs_links(ogc_server_settings.public_url + 'wcs?', instance.typename.encode('utf-8'),
             bbox=instance.bbox[:-1], crs=instance.bbox[-1], height=height, width=width)
         for ext, name, mime, wcs_url in links:
             Link.objects.get_or_create(resource= instance.resourcebase_ptr,
@@ -495,7 +503,7 @@ def geoserver_post_save(instance, sender, **kwargs):
         instance.set_gen_level(ANONYMOUS_USERS,permissions['anonymous'])
         instance.set_gen_level(AUTHENTICATED_USERS,permissions['authenticated'])
 
-    kml_reflector_link_download = ogc_server_settings.LOCATION + "wms/kml?" + urllib.urlencode({
+    kml_reflector_link_download = ogc_server_settings.public_url + "wms/kml?" + urllib.urlencode({
         'layers': instance.typename.encode('utf-8'),
         'mode': "download"
     })
@@ -510,7 +518,7 @@ def geoserver_post_save(instance, sender, **kwargs):
                         )
                     )
 
-    kml_reflector_link_view = ogc_server_settings.LOCATION + "wms/kml?" + urllib.urlencode({
+    kml_reflector_link_view = ogc_server_settings.public_url + "wms/kml?" + urllib.urlencode({
         'layers': instance.typename.encode('utf-8'),
         'mode': "refresh"
     })
@@ -525,7 +533,7 @@ def geoserver_post_save(instance, sender, **kwargs):
                         )
                     )
 
-    tile_url = ('%sgwc/service/gmaps?' % ogc_server_settings.LOCATION +
+    tile_url = ('%sgwc/service/gmaps?' % ogc_server_settings.public_url +
                 'layers=%s' % instance.typename.encode('utf-8') +
                 '&zoom={z}&x={x}&y={y}' +
                 '&format=image/png8'
@@ -557,8 +565,7 @@ def geoserver_post_save(instance, sender, **kwargs):
     #remove links that belong to and old address
 
     for link in instance.link_set.all():
-        if not urlparse(settings.SITEURL).hostname == urlparse(link.url).hostname and not \
-                    urlparse(ogc_server_settings.LOCATION).hostname == urlparse(link.url).hostname:
+        if not urlparse(ogc_server_settings.public_url).hostname == urlparse(link.url).hostname:
             link.delete()
 
     #Save layer attributes
@@ -581,6 +588,7 @@ def set_styles(layer, gs_catalog):
         style_set.append(save_style(alt_style))
 
     layer.styles = style_set
+    return layer
 
 def save_style(gs_style):
     style, created = Style.objects.get_or_create(name = gs_style.sld_name)
@@ -593,7 +601,7 @@ def save_style(gs_style):
 
 def is_layer_attribute_aggregable(store_type, field_name, field_type):
     """
-    Dechiper whether layer attribute is suitable for statistical derivation
+    Decipher whether layer attribute is suitable for statistical derivation
     """
 
     # must be vector layer
@@ -617,13 +625,15 @@ def get_attribute_statistics(layer_name, field):
 
     logger.debug('Deriving aggregate statistics for attribute %s', field)
 
+    if not ogc_server_settings.WPS_ENABLED:
+        return None
     try:
         return wps_execute_layer_attribute_statistics(layer_name, field)
     except Exception:
         logger.exception('Error generating layer aggregate statistics')
 
 
-def set_attributes(layer):
+def set_attributes(layer, overwrite=False):
     """
     Retrieve layer attribute names & types from Geoserver,
     then store in GeoNode database using Attribute model
@@ -647,7 +657,7 @@ def set_attributes(layer):
 
     attribute_map = []
     if layer.storeType == "dataStore":
-        dft_url = ogc_server_settings.LOCATION + "wfs?" + urllib.urlencode({
+        dft_url = ogc_server_settings.public_url + "wfs?" + urllib.urlencode({
             "service": "wfs",
             "version": "1.0.0",
             "request": "DescribeFeatureType",
@@ -661,7 +671,7 @@ def set_attributes(layer):
         except Exception:
             attribute_map = []
     elif layer.storeType == "coverageStore":
-        dc_url = ogc_server_settings.LOCATION + "wcs?" + urllib.urlencode({
+        dc_url = ogc_server_settings.public_url + "wcs?" + urllib.urlencode({
             "service": "wcs",
             "version": "1.1.0",
             "request": "DescribeCoverage",
@@ -682,7 +692,7 @@ def set_attributes(layer):
         for field, ftype in attribute_map:
             if field == la.attribute:
                 lafound = True
-        if not lafound:
+        if overwrite or not lafound:
             logger.debug("Going to delete [%s] for [%s]", la.attribute, layer.name.encode('utf-8'))
             la.delete()
